Redoing part of the domain of the application
