--- TeachTrack.Service/Domain/Application/Interfaces/IStudentsRepository.cs
@@ -1,8 +1,0 @@
-using TeachTrack.Service.Domain.Model.Students;
-
-namespace TeachTrack.Service.Domain.Application.Interfaces;
-
-public interface IStudentsRepository {
-   List<Student> Search(string schoolEnrollment);
-   
-}
