+++ TeachTrack.Service/Domain/Entities/Person/Person.cs
@@ -1,0 +1,11 @@
+namespace TeachTrack.Service.Domain.Entities.Person;
+
+public class Person {
+   public string Name { get; set; }
+   public string LastName { get; set; }
+
+   public Person(string name, string lastName) {
+      Name = name;
+      LastName = lastName;
+   }
+}
