+++ TeachTrack.Service/Domain/Entities/Student/IStudent.cs
@@ -1,0 +1,5 @@
+namespace TeachTrack.Service.Domain.Entities.Student;
+
+public interface IStudent {
+   bool CanRegisterInANewCourse(float score);
+}
