+++ TeachTrack.Service/Domain/Entities/Student/Student.cs
@@ -1,0 +1,16 @@
+using TeachTrack.Service.Domain.ValueObjects;
+
+namespace TeachTrack.Service.Domain.Entities.Student;
+
+public class Student : Person.Person, IStudent {
+   public string StudentId { get; set; }
+   public float StudentScore { get; set; }
+   private readonly Score _score;
+   
+   public Student(string studentId, string name, string lastName) : base(name, lastName) {
+      StudentId = studentId;
+   }
+   public bool CanRegisterInANewCourse(float score) {
+      return score <= _score.MaxScore && score >= _score.MinScore;
+   }
+}
