--- TeachTrack.Service/Domain/Model/Students/IStudent.cs
@@ -1,5 +1,0 @@
-namespace TeachTrack.Service.Domain.Model.Students;
-
-public interface IStudent {
-   bool CanRegisterNewSubjects(float score);
-}
