+++ TeachTrack.Service/Domain/ValueObjects/Score.cs
@@ -1,0 +1,6 @@
+namespace TeachTrack.Service.Domain.ValueObjects;
+
+public class Score {
+   public readonly float MaxScore = 250.00f;
+   public readonly float MinScore = 110.00f;
+}
