--- TeachTrack.Service/External/Enums/Score.cs
@@ -1,6 +1,0 @@
-namespace TeachTrack.Service.External.Enums;
-
-public class Score {
-   public  readonly float MinimumScore = 150;
-   public  readonly float MaximumScore = 270;
-}
