+++ TeachTrack.Service/TeachTrack.Service.csproj
@@ -7,7 +7,11 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <Folder Include="Domain\Application\" />
+      <PackageReference Include="Npgsql.ConfigureConnection" Version="0.1.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Application\" />
     </ItemGroup>
 
 </Project>
