+++ TeachTrack.Service/appsettings.Develop.json
@@ -1,0 +1,11 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=LocalHost;Port=5432;Database=TeachTrack;Username=hulara;Password=Descubrir2001.."
+  }
+}
